Refactor: deprecate tern/helpers folder
The tern/helpers directory contains common.py and docker.py. In an
effort to isolate docker functionality in to it's own module, this
commit moves tern/helpers/docker.py --> tern/analyze/docker/helpers.py.
It also moves tern/helpers/common.py --> tern/analyze/common.py. The
other changes in this commit update the import statements and
documentation to correctly reference the refactoring of the helpers
directory.

Works towards #262
