+++ tern/analyze/common.py
