+++ tern/analyze/docker/analyze.py
@@ -14,8 +14,8 @@ from tern.report import formats
 from tern.utils import constants
 from tern.utils import rootfs
 from tern.classes.notice import Notice
-from tern.helpers import common
-import tern.helpers.docker as dhelper
+from tern.analyze import common
+import tern.analyze.docker.helpers as dhelper
 from tern.command_lib import command_lib
