--- tern/helpers/__init__.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# Copyright (c) 2019 VMware, Inc. All Rights Reserved.
-# SPDX-License-Identifier: BSD-2-Clause
